requirements.txt
fastapi
uvicorn
pydantic
Pillow
python-multipart
python-dotenv
requests
json